--- src/Microsoft.Extensions.Options/IOptionsBuilder.cs
@@ -1,255 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace Microsoft.Extensions.Options
-{
-    /// <summary>
-    /// Used to configure TOptions instances.
-    /// </summary>
-    /// <typeparam name="TOptions">The type of options being requested.</typeparam>
-    public class OptionsBuilder<TOptions> where TOptions : class, new()
-    {
-        /// <summary>
-        /// The default name of the TOptions instance.
-        /// </summary>
-        public string Name { get; }
-
-        /// <summary>
-        /// The <see cref="IServiceCollection"/> for the options being configured.
-        /// </summary>
-        public IServiceCollection Services { get; }
-
-        /// <summary>
-        /// Constructor.
-        /// </summary>
-        /// <param name="services">The <see cref="IServiceCollection"/> for the options being configured.</param>
-        /// <param name="name">The default name of the TOptions instance, if null Options.DefaultName is used.</param>
-        public OptionsBuilder(IServiceCollection services, string name)
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            Services = services;
-            Name = name ?? Options.DefaultName;
-        }
-
-        public virtual OptionsBuilder<TOptions> Configure(Action<TOptions> configureOptions)
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddSingleton<IConfigureOptions<TOptions>>(new ConfigureNamedOptions<TOptions>(Name, configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> Configure<TDep>(Action<TOptions, TDep> configureOptions)
-            where TDep : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IConfigureOptions<TOptions>>(sp =>
-                new ConfigureNamedOptions<TOptions, TDep>(Name, sp.GetRequiredService<TDep>(), configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> Configure<TDep1, TDep2>(Action<TOptions, TDep1, TDep2> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IConfigureOptions<TOptions>>(sp =>
-                new ConfigureNamedOptions<TOptions, TDep1, TDep2>(Name, sp.GetRequiredService<TDep1>(), sp.GetRequiredService<TDep2>(), configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> Configure<TDep1, TDep2, TDep3>(Action<TOptions, TDep1, TDep2, TDep3> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-            where TDep3 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IConfigureOptions<TOptions>>(
-                sp => new ConfigureNamedOptions<TOptions, TDep1, TDep2, TDep3>(
-                    Name, 
-                    sp.GetRequiredService<TDep1>(), 
-                    sp.GetRequiredService<TDep2>(),
-                    sp.GetRequiredService<TDep3>(),
-                    configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> Configure<TDep1, TDep2, TDep3, TDep4>(Action<TOptions, TDep1, TDep2, TDep3, TDep4> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-            where TDep3 : class
-            where TDep4 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IConfigureOptions<TOptions>>(
-                sp => new ConfigureNamedOptions<TOptions, TDep1, TDep2, TDep3, TDep4>(
-                    Name,
-                    sp.GetRequiredService<TDep1>(),
-                    sp.GetRequiredService<TDep2>(),
-                    sp.GetRequiredService<TDep3>(),
-                    sp.GetRequiredService<TDep4>(),
-                    configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> Configure<TDep1, TDep2, TDep3, TDep4, TDep5>(Action<TOptions, TDep1, TDep2, TDep3, TDep4, TDep5> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-            where TDep3 : class
-            where TDep4 : class
-            where TDep5 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IConfigureOptions<TOptions>>(
-                sp => new ConfigureNamedOptions<TOptions, TDep1, TDep2, TDep3, TDep4, TDep5>(
-                    Name,
-                    sp.GetRequiredService<TDep1>(),
-                    sp.GetRequiredService<TDep2>(),
-                    sp.GetRequiredService<TDep3>(),
-                    sp.GetRequiredService<TDep4>(),
-                    sp.GetRequiredService<TDep5>(),
-                    configureOptions));
-            return this;
-        }
-
-        /// <summary>
-        /// Registers an action used to configure a particular type of options.
-        /// Note: These are run after all <seealso cref="Configure(Action{TOptions})"/>.
-        /// </summary>
-        /// <param name="configureOptions">The action used to configure the options.</param>
-        /// <returns>The <see cref="IServiceCollection"/> so that additional calls can be chained.</returns>
-        public virtual OptionsBuilder<TOptions> PostConfigure(Action<TOptions> configureOptions)
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddSingleton<IPostConfigureOptions<TOptions>>(new PostConfigureOptions<TOptions>(Name, configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> PostConfigure<TDep>(Action<TOptions, TDep> configureOptions)
-            where TDep : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IPostConfigureOptions<TOptions>>(sp =>
-                new PostConfigureOptions<TOptions, TDep>(Name, sp.GetRequiredService<TDep>(), configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> PostConfigure<TDep1, TDep2>(Action<TOptions, TDep1, TDep2> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IPostConfigureOptions<TOptions>>(sp =>
-                new PostConfigureOptions<TOptions, TDep1, TDep2>(Name, sp.GetRequiredService<TDep1>(), sp.GetRequiredService<TDep2>(), configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> PostConfigure<TDep1, TDep2, TDep3>(Action<TOptions, TDep1, TDep2, TDep3> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-            where TDep3 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IPostConfigureOptions<TOptions>>(
-                sp => new PostConfigureOptions<TOptions, TDep1, TDep2, TDep3>(
-                    Name,
-                    sp.GetRequiredService<TDep1>(),
-                    sp.GetRequiredService<TDep2>(),
-                    sp.GetRequiredService<TDep3>(),
-                    configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> PostConfigure<TDep1, TDep2, TDep3, TDep4>(Action<TOptions, TDep1, TDep2, TDep3, TDep4> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-            where TDep3 : class
-            where TDep4 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IPostConfigureOptions<TOptions>>(
-                sp => new PostConfigureOptions<TOptions, TDep1, TDep2, TDep3, TDep4>(
-                    Name,
-                    sp.GetRequiredService<TDep1>(),
-                    sp.GetRequiredService<TDep2>(),
-                    sp.GetRequiredService<TDep3>(),
-                    sp.GetRequiredService<TDep4>(),
-                    configureOptions));
-            return this;
-        }
-
-        public virtual OptionsBuilder<TOptions> PostConfigure<TDep1, TDep2, TDep3, TDep4, TDep5>(Action<TOptions, TDep1, TDep2, TDep3, TDep4, TDep5> configureOptions)
-            where TDep1 : class
-            where TDep2 : class
-            where TDep3 : class
-            where TDep4 : class
-            where TDep5 : class
-        {
-            if (configureOptions == null)
-            {
-                throw new ArgumentNullException(nameof(configureOptions));
-            }
-
-            Services.AddTransient<IPostConfigureOptions<TOptions>>(
-                sp => new PostConfigureOptions<TOptions, TDep1, TDep2, TDep3, TDep4, TDep5>(
-                    Name,
-                    sp.GetRequiredService<TDep1>(),
-                    sp.GetRequiredService<TDep2>(),
-                    sp.GetRequiredService<TDep3>(),
-                    sp.GetRequiredService<TDep4>(),
-                    sp.GetRequiredService<TDep5>(),
-                    configureOptions));
-            return this;
-        }
-    }
-}
