+++ test/Microsoft.Extensions.Options.Test/OptionsBuilderTest.cs
@@ -79,5 +79,112 @@ namespace Microsoft.Extensions.Options.Tests
             var factory = sp.GetRequiredService<IOptionsFactory<FakeOptions>>();
             Assert.Equal("ABCDE", factory.Create("1").Message);
         }
+
+
+        public class SomeService
+        {
+            public SomeService(string stuff) => Stuff = stuff;
+
+            public string Stuff { get; set; }
+        }
+
+        public class Counter
+        {
+            public int Count { get; set; }
+
+            public void Increment() => Count++;
+        }
+
+        public class SomeCounterConsumer
+        {
+            public SomeCounterConsumer(Counter c)
+            {
+                Current = c.Count;
+                c.Increment();
+            }
+
+            public int Current { get; }
+        }
+
+        [Fact]
+        public void ConfigureOptionsWithSingletonDepWillUpdate()
+        {
+            var someService = new SomeService("Something");
+            var services = new ServiceCollection().AddOptions().AddSingleton(someService);
+            services.AddOptions<FakeOptions>().Configure<SomeService>((o, s) => o.Message = s.Stuff);
+            services.AddOptions<FakeOptions>("named").Configure<SomeService>((o, s) => o.Message = "named " + s.Stuff);
+
+            var sp = services.BuildServiceProvider();
+            var factory = sp.GetRequiredService<IOptionsFactory<FakeOptions>>();
+
+            Assert.Equal("named Something", factory.Create("named").Message);
+            Assert.Equal("Something", factory.Create(Options.DefaultName).Message);
+
+            someService.Stuff = "Else";
+
+            Assert.Equal("named Else", factory.Create("named").Message);
+            Assert.Equal("Else", factory.Create(Options.DefaultName).Message);
+        }
+
+        [Fact]
+        public void ConfigureOptionsWithTransientDep()
+        {
+            var services = new ServiceCollection().AddOptions()
+                .AddSingleton<Counter>()
+                .AddTransient<SomeCounterConsumer>();
+            services.AddOptions<FakeOptions>().Configure(o => o.Message = "none");
+            services.AddOptions<FakeOptions>("1dep").Configure<SomeCounterConsumer>((o, s) => o.Message = s.Current + "");
+            services.AddOptions<FakeOptions>("2dep").Configure<SomeCounterConsumer, SomeCounterConsumer>((o, s, s2) => o.Message = s.Current + " " + s2.Current);
+            services.AddOptions<FakeOptions>("3dep").Configure<SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer>((o, s, s2, s3) => o.Message = s.Current + " " + s2.Current + " " + s3.Current);
+            services.AddOptions<FakeOptions>("4dep").Configure<SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer>((o, s, s2, s3, s4) => o.Message = s.Current + " " + s2.Current + " " + s3.Current + " " + s4.Current);
+            services.AddOptions<FakeOptions>("5dep").Configure<SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer>((o, s, s2, s3, s4, s5) => o.Message = s.Current + " " + s2.Current + " " + s3.Current + " " + s4.Current + " " + s5.Current);
+
+            var sp = services.BuildServiceProvider();
+            var factory = sp.GetRequiredService<IOptionsFactory<FakeOptions>>();
+
+            Assert.Equal("none", factory.Create(Options.DefaultName).Message);
+            Assert.Equal("0", factory.Create("1dep").Message);
+            Assert.Equal("1 2", factory.Create("2dep").Message);
+            Assert.Equal("3 4 5", factory.Create("3dep").Message);
+            Assert.Equal("6 7 8 9", factory.Create("4dep").Message);
+            Assert.Equal("10 11 12 13 14", factory.Create("5dep").Message);
+
+            // Factory caches configures
+            Assert.Equal("0", factory.Create("1dep").Message);
+
+            // New factory will reexecute
+            Assert.Equal("15", sp.GetRequiredService<IOptionsFactory<FakeOptions>>().Create("1dep").Message);
+        }
+
+        [Fact]
+        public void PostConfigureOptionsWithTransientDep()
+        {
+            var services = new ServiceCollection().AddOptions()
+                .AddSingleton<Counter>()
+                .AddTransient<SomeCounterConsumer>();
+            services.ConfigureAll<FakeOptions>(o => o.Message = "Override");
+            services.AddOptions<FakeOptions>().PostConfigure(o => o.Message = "none");
+            services.AddOptions<FakeOptions>("1dep").PostConfigure<SomeCounterConsumer>((o, s) => o.Message = s.Current + "");
+            services.AddOptions<FakeOptions>("2dep").PostConfigure<SomeCounterConsumer, SomeCounterConsumer>((o, s, s2) => o.Message = s.Current + " " + s2.Current);
+            services.AddOptions<FakeOptions>("3dep").PostConfigure<SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer>((o, s, s2, s3) => o.Message = s.Current + " " + s2.Current + " " + s3.Current);
+            services.AddOptions<FakeOptions>("4dep").PostConfigure<SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer>((o, s, s2, s3, s4) => o.Message = s.Current + " " + s2.Current + " " + s3.Current + " " + s4.Current);
+            services.AddOptions<FakeOptions>("5dep").PostConfigure<SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer, SomeCounterConsumer>((o, s, s2, s3, s4, s5) => o.Message = s.Current + " " + s2.Current + " " + s3.Current + " " + s4.Current + " " + s5.Current);
+
+            var sp = services.BuildServiceProvider();
+            var factory = sp.GetRequiredService<IOptionsFactory<FakeOptions>>();
+
+            Assert.Equal("none", factory.Create(Options.DefaultName).Message);
+            Assert.Equal("0", factory.Create("1dep").Message);
+            Assert.Equal("1 2", factory.Create("2dep").Message);
+            Assert.Equal("3 4 5", factory.Create("3dep").Message);
+            Assert.Equal("6 7 8 9", factory.Create("4dep").Message);
+            Assert.Equal("10 11 12 13 14", factory.Create("5dep").Message);
+
+            // Factory caches configures
+            Assert.Equal("0", factory.Create("1dep").Message);
+
+            // New factory will reexecute
+            Assert.Equal("15", sp.GetRequiredService<IOptionsFactory<FakeOptions>>().Create("1dep").Message);
+        }
     }
 }
